Move `OtherState` to its own file.
